core: Refactoring POSIX-only code for build tags
